vanilla/requirements.txt
fastapi
pydantic
pydantic[email]
psycopg>=3.2.9
passlib
python-jose
uvicorn